updated Context, made alpha float, some other clean ups, compiles

+++ sparrow-sharp/Sparrow/Core/RenderState.cs
@@ -4,28 +4,28 @@ using Sparrow.Geom;
 
 namespace Sparrow.Core
 {
-	public class RenderState
-	{
-		public Matrix ModelViewMatrix;
-		public double Alpha;
-		public uint BlendMode;
+    public class RenderState
+    {
+        public Matrix ModelViewMatrix;
+        public float Alpha;
+        public uint BlendMode;
 
-		public RenderState ()
-		{
-			ModelViewMatrix = new Matrix ();
-			Alpha = 1.0;
-			BlendMode = Sparrow.Display.BlendMode.NORMAL;
-		}
+        public RenderState()
+        {
+            ModelViewMatrix = new Matrix();
+            Alpha = 1.0f;
+            BlendMode = Sparrow.Display.BlendMode.NORMAL;
+        }
 
-		public void Setup (RenderState state, Matrix modelViewMatrix, double alpha, uint blendMode)
-		{
-			Alpha = alpha * state.Alpha;
-			BlendMode = blendMode == Sparrow.Display.BlendMode.AUTO ? state.BlendMode : blendMode;
+        public void Setup(RenderState state, Matrix modelViewMatrix, float alpha, uint blendMode)
+        {
+            Alpha = alpha * state.Alpha;
+            BlendMode = blendMode == Sparrow.Display.BlendMode.AUTO ? state.BlendMode : blendMode;
 
-			ModelViewMatrix = new Matrix ();
-			ModelViewMatrix.CopyFromMatrix (state.ModelViewMatrix);
-			ModelViewMatrix.PrependMatrix (modelViewMatrix);
-		}
-	}
+            ModelViewMatrix = new Matrix();
+            ModelViewMatrix.CopyFromMatrix(state.ModelViewMatrix);
+            ModelViewMatrix.PrependMatrix(modelViewMatrix);
+        }
+    }
 }
